learnflow_app/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
openai==1.3.0
intasend==1.0.0
psycopg2-binary==2.9.7  # For PostgreSQL